Refactor Manage Products page and enhance Home page with product fetching and loading states
